smbm: lazy-evaluated smbool
Added support for lazy evaluation to SMBool.  This allows knows/items to
be formed as a list-of-lists, then flattened at the end.

For now this is a significant reduction in performance, but this will
enable big gains when the code is modified to not use knows/items until
the end.