cache: Make Cache an instance
Changed Cache to be a singleton instance of VersionedCache rather than
just using the Cache class as a namespace.  This slightly improves
performance, as we no longer have to do a name lookup to access Cache
from within the cache itself.  As a bonus, it makes it simpler to write
unit tests, if that is ever desired.