cache: Comment-out debug code
Commented-out the code that validates the contents of the cache.  It's
a shame that this affects performance (not much, but it's measurable).
The rest of the cache debug logging can still be enabled via the
command-line, as it is out of the fast path.

It is possible to be able to dynamically enable the flag without
affecting performaince via metaprogramming techinques, but this would
affect readability of the code and is probably not worth the time except
as a programming challenge.