cache: Use list instead of dict
Changed the cache to use a list instead of a dict for storing the
results of functions.  This eliminates a dictionary lookup, as each
decorator gets its own slot at a specific index in the list, at the cost
of having to construct the list each time we change the cache key.  This
provides a small but measurable speed improvement over using a dict.