smbool: Use smboolFalse in wnot
Modified wnot to return smboolFalse when the argument is truthy; this is
faster than constructing a new SMBool.

This is a change in behavior: the old behavior was to construct a falsy
SMBool with the same difficulty as the boolean being negated; the new
behavior is more consistent with wand/wor, which use smboolFalse.