smbm: Rename evalWithItem to withItem
Renamed evalWithItem to just withItem and changed the semantics
slightly.  It didn't make sense to have evalWithItem any more, now that
eval has been removed.

Instead, withItem takes an item and a function (evalWithItem took them
in the opposite order) and calls with function (but without the
SMBoolManager as a parameter).  This does not affect performance much,
as this path is not called often (at least not for a typical SMRAT
seed).