smbool: Add wandmax
Added a new function wandmax to SMBool, which is like wand, but takes
the maximum instead of the sum.

Modified graph and helpers to use wandmax instead of custom one-off
code.  Since wandmax has been optimized, this gives a small performance
boost, despite the extra overhead of the function call.