Fix objectives data crash when starting autotracking
This crash occurs when starting autotracking and reading objective data from the ROM.  I'm not sure why the state data is not correct here (it seems to be missing the objective data) but it may be related to some difference between the Archipelago ROM vs the VARIA ROM changes.  This at least prevents that crash and lets autotracking work correctly for AP ROMs again.